pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage-env"
version = "1.0.0"
description = "Email Triage Environment for OpenEnv — classify, respond, and route emails"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "email_triage_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["email_triage_env", "email_triage_env.server"]
package-dir = { "email_triage_env" = "email_triage_env", "email_triage_env.server" = "email_triage_env/server" }
